chore(gitignore): add dist and .vscode to gitignore
This commit adds the `dist` and `.vscode` directories to the `.gitignore` file. This prevents these directories from being committed to the repository.

The `dist` directory contains the compiled output of the project, which is not necessary to track in version control. The `.vscode` directory contains configuration files for the Visual Studio Code editor, which are also not necessary to track in version control.